Refactor client code to extract actions that will need be to reused
These actions will be reused in the aside controller, we move them
to a common file than can be imported by both features.
